requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
boto3==1.26.0
requests==2.31.0
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart
NLTK
sentence-transformers
spacy
transformers
sentence_transformers
nltk
openai
spacy==3.5.0